LayoutTests/svg/animations/svgenum-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test SVGUnitTypes enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pattern.patternContentUnits.animVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.animVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.animVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS pattern.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.animVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS pattern.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS pattern.patternContentUnits.animVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS pattern.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS successfullyParsed is true

TEST COMPLETE
